feat(SPV-742): fix linter errors

+++ transports/p2p/connmgr/seed.go
@@ -68,7 +68,7 @@ func SeedFromDNS(chainParams *chaincfg.Params, reqServices wire.ServiceFlag,
 					// and 7 days ago.
 					time.Now().Add(-1*time.Second*time.Duration(secondsIn3Days+
 						randSource.Int31n(secondsIn4Days))), //nolint:gosec
-					0, peer, uint16(intPort))
+					0, peer, intPort)
 			}
 
 			seedFn(addresses)
